fix imports and other basic problems

+++ comfy/app/user_manager.py
@@ -22,7 +22,7 @@ default_user = "default"
 class FileInfo(TypedDict):
     path: str
     size: int
-    modified: int
+    modified: float
 
 
 def get_file_info(path: str, relative_to: str) -> FileInfo:
